test: renamed manager test file and updated test suite
